[frontend/backend] share a common users list component
