+++ opencti-platform/opencti-graphql/config/schema/opencti.graphql
@@ -1098,7 +1098,7 @@ input UsersFiltering {
 }
 type UserConnection {
   pageInfo: PageInfo!
-  edges: [UserEdge]
+  edges: [UserEdge!]!
 }
 type CreatorConnection {
   pageInfo: PageInfo!
